requirements.txt
# Core dependencies
requests>=2.28.0
beautifulsoup4>=4.11.0
PyPDF2>=3.0.0
pyyaml>=6.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Optional but recommended
lxml>=4.9.0
html5lib>=1.1